requirements.txt
pandas==1.5.2
PyQt6==6.4.0
pillow==9.4.0
pylatex==1.4.2
